backend/pyproject.toml
[tool.poetry]
name = "blue_light-api"
version = "0.1.0"
description = ""
repository = "https://code.bld.ai/house-numoo/templates/django-react"
authors = ["Ahmed Mohamedeen <ah.mohamadeen@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "*"
django-environ = "^0.9.0"
celery = "^5.2.7"
psycopg2 = "^2.9.3"
redis = "^4.3.4"
django-celery-beat = "^2.3.0"
django-celery-results = "^2.4.0"
uvicorn = "^0.18.3"
django-cors-headers = "^3.13.0"
django-rest-knox = "^4.2.0"
djangorestframework = "^3.15.1"
django-admin-rangefilter = "^0.12.4"
pillow = "^10.4.0"
django-nested-inline = "^0.4.6"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
setuptools = "^74.1.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-modeltranslation = "^0.19.12"
django-safedelete = "^1.4.0"
django-simple-history = "^3.7.0"
django-redis = "^5.4.0"
phonenumbers = "^8.13.52"
uritemplate = "^4.1.1"
urllib3 = "^2.3.0"
tzdata = "^2024.2"
typing-extensions = "^4.12.2"
django-debug-toolbar = "^4.4.6"
charset-normalizer = "^3.4.1"
cachetools = "^5.5.0"
cachecontrol = "^0.14.1"

[tool.poetry.group.dev.dependencies]
pylint = "2.15.3"
black = "22.8.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
coverage = "^6.5.0"
safety = "^2.2.0"
django-coverage-plugin = "^2.0.3"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
ensure_newline_before_comments = false
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.mypy_cache
  | \.venv
  | dist
  | migrations
)/
'''

[tool.isort]
profile = 'black'
